Move health check out of transaction scope
* Break UpdateStrategy into additional steps

Fixes #280
